Add GameManager, responsible for creating a Game
The GameManager will be created on server startup. The Clue-Less "Home
Page" will access the GameManager and add client's name and chosen
suspect. The GameManager allows a Game to be created. The GameManager
might need additional work: it might be good for him to be able to answer
whether or not a particular client (via his name) is currently in a game,
and if so, which one.