Changed Game's State to a Queue of State
Game now holds a Queue of state for each Player. The head is the oldest
state, the tail is the newest. As the PollServlet is called, it grabs the
oldest state. As Actions are performed, they use the newest state to
create the next newest state. This was done so that States are not lost
when they're updated faster than the PollServlet happens to see them. The
Game manages removing old states after they've been seen.

Fixed typo in State.Builder method.