Using game.html instead of game.jsp/play.jsp
Using game.html will make debugging formatting faster because we will not
have to run the server each time we want to check something (just open the
file in a browser).

InitializeGameServlet -- sends a JSON object that is used to populate the
game.html page when users first start the game.